requirements.txt
# Edge Reliability Gap — Python Dependencies
# Install:  pip install -r requirements.txt
#
# GPU inference requires CUDA-compatible PyTorch.
# See https://pytorch.org/get-started/locally/ for install instructions.

# ── Core ML stack ──────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.25.0
bitsandbytes>=0.41.0          # NF4 quantisation for Qwen2.5-VL-7B

# ── Model-specific ────────────────────────────────────────────
qwen-vl-utils>=0.0.2          # Image pre-processing for Qwen2.5-VL

# ── Data ──────────────────────────────────────────────────────
datasets>=2.16.0              # HuggingFace dataset streaming
Pillow>=9.0.0
requests>=2.25.0

# ── Analysis & visualisation ──────────────────────────────────
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.5.0
scikit-learn>=1.2.0           # Cohen's kappa

# ── LLM-as-Judge (Phase 5 only — optional) ───────────────────
openai>=1.0.0                 # GPT-4o API; only needed for llm_judge.py
